chore: move app for apps/examples
