+++ apps/examples/.prettierrc.js
