+++ apps/examples/package.json
@@ -7,14 +7,14 @@
     "check-types": "tsc --noEmit"
   },
   "dependencies": {
-    "@react-navigation/bottom-tabs": "link:../react-navigation/packages/bottom-tabs/",
-    "@react-navigation/core": "link:../react-navigation/packages/core/",
-    "@react-navigation/drawer": "link:../react-navigation/packages/drawer/",
-    "@react-navigation/elements": "link:../react-navigation/packages/elements/",
-    "@react-navigation/native": "link:../react-navigation/packages/native/",
-    "@react-navigation/native-stack": "link:../react-navigation/packages/native-stack/",
-    "@react-navigation/routers": "link:../react-navigation/packages/routers/",
-    "@react-navigation/stack": "link:../react-navigation/packages/stack/",
+    "@react-navigation/bottom-tabs": "link:../../react-navigation/packages/bottom-tabs/",
+    "@react-navigation/core": "link:../../react-navigation/packages/core/",
+    "@react-navigation/drawer": "link:../../react-navigation/packages/drawer/",
+    "@react-navigation/elements": "link:../../react-navigation/packages/elements/",
+    "@react-navigation/native": "link:../../react-navigation/packages/native/",
+    "@react-navigation/native-stack": "link:../../react-navigation/packages/native-stack/",
+    "@react-navigation/routers": "link:../../react-navigation/packages/routers/",
+    "@react-navigation/stack": "link:../../react-navigation/packages/stack/",
     "nanoid": "^4.0.2",
     "react": "18.2.0",
     "react-native": "0.74.1",
@@ -22,11 +22,11 @@
     "react-native-reanimated": "3.11.0",
     "react-native-restart": "^0.0.27",
     "react-native-safe-area-context": "^4.10.1",
-    "react-native-screens": "link:../",
+    "react-native-screens": "link:../../",
     "react-native-vector-icons": "^8.0.0"
   },
   "resolutions": {
-    "@react-navigation/core": "link:../react-navigation/packages/core/"
+    "@react-navigation/core": "link:../../react-navigation/packages/core/"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
