+++ apps/examples/src/assets/trees.jpg
