+++ apps/examples/src/screens/BottomTabsAndStack.tsx
