+++ apps/examples/src/screens/SearchBar.tsx
