+++ apps/examples/src/screens/StackPresentation.tsx
