+++ apps/examples/src/screens/StatusBar.tsx
