+++ apps/examples/src/shared/SettingsInput.tsx
