+++ apps/examples/src/shared/SettingsPicker.tsx
