+++ apps/examples/src/shared/SettingsSwitch.tsx
