+++ apps/examples/src/shared/Snack.tsx
