+++ apps/examples/src/shared/index.ts
